backend/requirements.txt
fastapi
uvicorn
pydantic-settings
pymongo
langchain
PyMuPDF
python-multipart
google-generativeai
